+++ shardingsphere-underlying/shardingsphere-route/src/main/java/org/apache/shardingsphere/underlying/route/DateNodeRouteDecorator.java
@@ -15,9 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.route.router;
-
-import org.apache.shardingsphere.underlying.route.RouteContext;
+package org.apache.shardingsphere.underlying.route;
 
 /**
  * Data node route decorator.
